"port" to wasm32. This turned out to be straightforward.

+++ arch/wasm32/ksigaction.h
@@ -1,6 +1,6 @@
 struct k_sigaction {
 	void (*handler)(int);
-	unsigned long flags;
+	unsigned long long flags;
 	void (*restorer)(void);
 	unsigned mask[2];
 };
